Changed float price to double

+++ src/businessLogic/ApplicationFacadeImpl.java
@@ -50,7 +50,7 @@ public final class ApplicationFacadeImpl  implements ApplicationFacadeInterface 
 		return ruralHouse;
 	}
 
-	public Offer createOffer(RuralHouse ruralHouse, Date firstDay, Date lastDay, float price) throws OverlappingOfferException, BadDatesException {
+	public Offer createOffer(RuralHouse ruralHouse, Date firstDay, Date lastDay, double price) throws OverlappingOfferException, BadDatesException {
 		System.out.println(">> ApplicationFacadeImpl: createOffer=> ruralHouse= "+ruralHouse+" firstDay= "+firstDay+" lastDay="+lastDay+" price="+price);
 
 		Offer offer = null;
